Fix memory violation: shared pointer passed to logging thread
logging_thread runs in a thread and is passed a Channel* as a void*
argument via pthread_create. Then, a shared_ptr takes over the Channel*
via reset() in logging_thread. This effectively gives ownership to the
shared_ptr.
This is incorrect, because the Channel is owned by another shared_ptr
outside of the thread. Passing a raw Channel* breaks the shared_ptr
semantics.

When the logging_thread ends, the shared_ptr deletes the Channel.
However, outside of logging_thread it is still used by a seperate
shared_ptr group. This causes crashes, double delete...
(It can be observed as crash on shutdown and with valgrind.)

Work around this problem by holding a shared_ptr to represent a link to
in the logging thread. Using std::thread would be nicer, but would
require a bigger patch.