add BUILD_TEST option for disabling unit-tests
During cross-compiling with some buildsystem (buildroot for example) we
have no possibility to run the implemented unit tests. This well known
fact is handled by the build system by passing several variables to the
running build, one of these is the "BUILD_TEST".

We use this for enabling/disabling building the unit tests.

Per default the BUILD_TEST is on, so no functional change to previous
versions is made.

Signed-off-by: Hannes Schmelzer <oe5hpm@oevsv.at>