Added automatic reopen of underlying device in case of libsml read 0. (#370)
* Added automatic reopen of underlying device in case of libsml read 0.

As a workaround for #362 added an automatic reopen if
sml_transport_read returns 0 bytes.
Issue: #362

* Fix GTEST_LIBS_DIR for gmock.

Don't really know why this has changed. Might be that
gmock changed it.

* Added reopen as well to read to prevent use of invalid fd.

If reopen fails the fd might be invalid and the next read call
does a write or pass an invalid fd to sml_transport_read.
Now reopen gets triggered again and if it fails a sleep of 1s
is done to prevent busy loopings.