remove "daemon" setting,
setting daemon:false was NEVER a useful mode for running vzlogger.
it NOT only disabled unix-"daemonization",
but it made the reading thread exit after getting a single reading,
while it did NOT NOT have vzlogger wait for the logging thread to actually
report the data to the api, so data would have been lost most cases.

to be replaced with one-shot and foreground options.
for now:
always run logging thread,
always daemonize.