fix combined use of `interval` and aggregation
it seems combined use of `interval` and aggregation was never considered.
the expected use-case of aggregation is to reduce the (externally determined)
rate at which a meter is generating data.
but some users want to use it create averages over data collected at a
configured `interval` - so let's support that.