remove meter_details_t.periodic,
this was supposed to indicate if a protocol
requires the reading delay/"period" to be define externally (via `interval`),
or if it limits the reading rate by itself.

this has been useless for a long time, because for many meters it depends
on the runtime configuration,
and actually a method to evaluate this at runtime has been introduced
as `allowInterval()` in cf3b676a1, many years ago.