reject readings with really old timestamps, ...
these typically occur when a meter is sending timestamps but does
not have it's clock set properly, typically returning seconds since
power-up instead of seconds since the unix epoch (1970-01-01).
this often causes confusion, which can be easily avoided.

i arbitrarily picked 1990-01-01 as a cut-off point,
for one, few or no smart meters should have existed back then,
so nobody will be trying to read archived data that old,
for another a meter counting it's uptime would need twenty
years of uptime to generate a timestamp that is accepted.

for any obscure corner cases, a config option:
	"i_have_a_time_machine"
is provided to disable the check.